% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_browse.r
\name{gh_browse}
\alias{gh_browse}
\title{Open the GitHub repository in a web browser}
\usage{
gh_browse(what = NULL, ..., branch = NULL, repo = NULL)
}
\arguments{
\item{what}{A character string, integer, or \code{NULL}
\itemize{
\item \code{NULL} opens up repository home page
\item "commit" opens up last commit
\item "projects" opens repository projects
\item "settings" opens repository settings
\item "wiki" opens repository wiki
\item INT or "INT" opens issue or pull request INT
\item "PATH" opens file page of file with path PATH
\item "PATH:INT" opens file page of file with path PATH at line INT
\item "SHA" opens commit page of commit with sha SHA
}}

\item{...}{Ignored.}

\item{branch}{A string of branch name to use (if other than default branch).}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_browse()} opens up a GitHub repository in a web browser.
}
\details{
To configure a web browser other than the default use
the \code{BROWSER} environmental variable.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_browse() # open up repo home page
  gh_browse(23) # issue or pull request 23
  gh_browse("settings") # open up repo settings
  gh_browse("R/utils.r:5") # open up file page for R/utils.r at line 5
}
}
\seealso{
\url{https://cli.github.com/manual/gh_browse}
}
