% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issue.r
\name{gh_issue_create}
\alias{gh_issue_create}
\title{Create issue in a repository}
\usage{
gh_issue_create(
  ...,
  assignee = NULL,
  body = "",
  label = NULL,
  milestone = NULL,
  project = NULL,
  title = NULL,
  repo = NULL
)
}
\arguments{
\item{...}{Ignored.}

\item{assignee}{Character vector or \code{NULL}.  Use \code{"@me"} to self-assign.}

\item{body}{String.}

\item{label}{Character vector or \code{NULL}.}

\item{milestone}{Character vector or \code{NULL}.}

\item{project}{Character vector or \code{NULL}.}

\item{title}{String.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly
}
\description{
\code{gh_issue_create()} creates an issue in a repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_issue_create(title = "New issue title")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_issue_create}
}
