% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_label.r
\name{gh_label_clone}
\alias{gh_label_clone}
\title{Clone labels from one repo to another}
\usage{
gh_label_clone(source_repository, ..., force = TRUE, repo = NULL)
}
\arguments{
\item{source_repository}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}

\item{...}{Ignored.}

\item{force}{If \code{TRUE} update any existing label(s).}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly
}
\description{
\code{gh_label_clone()} copies labels from one repo to another.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_label_clone("r-lib/cli") # copy labels from {ghcli} to this repo
}
}
\seealso{
\url{https://cli.github.com/manual/gh_label_clone}
}
