% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_label.r
\name{gh_label_create}
\alias{gh_label_create}
\title{Create label in a repository}
\usage{
gh_label_create(
  name,
  ...,
  color = NULL,
  description = NULL,
  force = TRUE,
  repo = NULL
)
}
\arguments{
\item{name}{String of label to create.}

\item{...}{Ignored.}

\item{color}{Color of label (default will be a random color).}

\item{description}{Description of label.}

\item{force}{If \code{TRUE} update an existing label.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly
}
\description{
\code{gh_label_create()} creates a label in a repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_label_create("upkeep", color = "grey", description = "Package maintenance")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_label_create}
}
