% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_label.r
\name{gh_label_delete}
\alias{gh_label_delete}
\title{Delete label(s) from a repository}
\usage{
gh_label_delete(name, ..., repo = NULL)
}
\arguments{
\item{name}{Character vector of label(s) to delete.}

\item{...}{Ignored.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly
}
\description{
\code{gh_label_delete()} deletes label(s) from a repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_label_delete("wontfix") # delete `wontfix` label
}
}
\seealso{
\url{https://cli.github.com/manual/gh_label_delete}
}
