% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_label.r
\name{gh_label_edit}
\alias{gh_label_edit}
\title{Update label in a repository}
\usage{
gh_label_edit(
  to_edit,
  ...,
  color = NULL,
  description = NULL,
  name = NULL,
  repo = NULL
)
}
\arguments{
\item{to_edit}{Name of the label to edit.}

\item{...}{Ignored.}

\item{color}{Color of label.}

\item{description}{Description of label.}

\item{name}{New name of the label.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly
}
\description{
\code{gh_label_edit()} updates a label in a repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_label_edit("question", color = "orange")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_label_edit}
}
