% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_label.r
\name{gh_label_list}
\alias{gh_label_list}
\title{List labels in GitHub repository}
\usage{
gh_label_list(..., repo = NULL)
}
\arguments{
\item{...}{Ignored.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
A data frame with labels with columns "name", "color", "description".
}
\description{
\code{gh_label_list()} returns a data frame of labels in a GitHub repository.
}
\details{
The "color" column in the returned data frame may have SGR control sequences in it.
They can be stripped by \code{df$color <- cli::ansi_strip(df$color)}.
Can also disable colors with a \code{NO_COLOR=true} environmental variable
or set \code{options(cli.ansi = FALSE)}.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_label_list()
}
}
