% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_archive}
\alias{gh_repo_archive}
\title{Archive a GitHub repository}
\usage{
gh_repo_archive(repository = NULL)
}
\arguments{
\item{repository}{Repository to archive.  If \code{NULL} archive current repository.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_archive()} archives a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_archive() # archive current repo
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_archive}
}
