% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_create}
\alias{gh_repo_create}
\title{Create a GitHub repository}
\usage{
gh_repo_create(
  repository,
  visibility = "private",
  ...,
  description = NULL,
  enable_issues = TRUE,
  enable_wiki = FALSE,
  homepage = NULL
)
}
\arguments{
\item{repository}{Name of repository in \verb{[OWNER/]REPO} format.  \code{OWNER} defaults to name of authenticating user.}

\item{visibility}{Either \code{"public"}, \code{"private"}, \code{"internal"}}

\item{...}{Ignored}

\item{description}{A string.}

\item{enable_issues}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_wiki}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{homepage}{A string (of an URL).}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_create()} creates a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_create("newrepo", "private", description = "A new GitHub repository")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_create}
}
