% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_delete}
\alias{gh_repo_delete}
\title{Delete a GitHub repository}
\usage{
gh_repo_delete(repository)
}
\arguments{
\item{repository}{Character vector of repositories in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_delete()} deletes a repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_delete("trevorld/repo_that_does_not_exist")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_delete}

\url{https://cli.github.com/manual/gh_label_delete}
}
