% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_edit}
\alias{gh_repo_edit}
\title{Edit repository settings.}
\usage{
gh_repo_edit(
  repository = NULL,
  ...,
  add_topic = NULL,
  allow_forking = NULL,
  allow_update_branch = NULL,
  default_branch = NULL,
  delete_branch_on_merge = NULL,
  description = NULL,
  enable_advanced_security = NULL,
  enable_discussions = NULL,
  enable_issues = NULL,
  enable_merge_commit = NULL,
  enable_projects = NULL,
  enable_rebase_merge = NULL,
  enable_secret_scanning = NULL,
  enable_secret_scanning_push_protection = NULL,
  enable_squash_merge = NULL,
  enable_wiki = NULL,
  homepage = NULL,
  remove_topic = NULL,
  template = NULL,
  visibility = NULL
)
}
\arguments{
\item{repository}{Repository in \verb{[HOST/]OWNER/REPO} format.
Default is the repository the working directory is in.}

\item{...}{Ignored}

\item{add_topic}{Either \code{NULL} or a character vector.}

\item{allow_forking}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{allow_update_branch}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{default_branch}{A string.}

\item{delete_branch_on_merge}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{description}{A string.}

\item{enable_advanced_security}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_discussions}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_issues}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_merge_commit}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_projects}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_rebase_merge}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_secret_scanning}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_secret_scanning_push_protection}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_squash_merge}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{enable_wiki}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{homepage}{A string (of an URL).}

\item{remove_topic}{Either \code{NULL} or a character vector.}

\item{template}{Either \code{NULL}, \code{TRUE}, or \code{FALSE}.}

\item{visibility}{Either \code{NULL}, \code{"public"}, \code{"private"}, \code{"internal"}.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_edit()} edits repository settings.
For all arguments \code{NULL} means don't edit this setting.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_edit(visibility = "public") # make a private repo public
  gh_repo_edit(delete_branch_on_merge = TRUE) # enable delete head branch when PRs are merged
  gh_repo_edit(enable_merge_commit = FALSE) # disable merging PRs with merge commits
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_edit}
}
