% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_list}
\alias{gh_repo_list}
\title{List GitHub repositories}
\usage{
gh_repo_list(
  ...,
  archived = FALSE,
  fork = FALSE,
  language = NULL,
  limit = 1000L,
  fields = c("nameWithOwner", "visibility", "stargazerCount", "description"),
  omit_fork = FALSE,
  omit_archived = FALSE,
  topic = NULL,
  visibility = NULL
)
}
\arguments{
\item{...}{Ignored}

\item{archived}{Show only archived repositories}

\item{fork}{Show only forks}

\item{language}{Filter by primary coding language}

\item{limit}{Maximum number of repositories to list}

\item{fields}{Data fields to include.  See \url{https://cli.github.com/manual/gh_repo_list}.}

\item{omit_fork}{Omit forks}

\item{omit_archived}{Omit archived repositories}

\item{topic}{Filter by topic}

\item{visibility}{Filter by repository visibility.  Either \code{NULL}, \code{"public"}, \code{"private"}, \code{"internal"}}
}
\description{
\code{gh_repo_list()} lists GitHub repositories
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  df <- gh_repo_list("new_name") # rename repo "new_name"
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_list}
}
