% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_rename}
\alias{gh_repo_rename}
\title{Rename a GitHub repository}
\usage{
gh_repo_rename(new_name, ..., repo = NULL)
}
\arguments{
\item{new_name}{New repository name.}

\item{...}{Ignored.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_rename()} renames a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_rename("new_name") # rename repo "new_name"
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_rename}
}
