% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_unarchive}
\alias{gh_repo_unarchive}
\title{Unarchive a GitHub repository}
\usage{
gh_repo_unarchive(repository = NULL)
}
\arguments{
\item{repository}{Repository to unarchive.  If \code{NULL} archive current repository.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_unarchive()} unarchives a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_unarchive() # unarchive current repo (if archived)
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_unarchive}
}
