% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo.r
\name{gh_repo_view}
\alias{gh_repo_view}
\title{View details of a GitHub repository}
\usage{
gh_repo_view(
  repository = NULL,
  ...,
  fields = c("nameWithOwner", "deleteBranchOnMerge", "description",
    "hasDiscussionsEnabled", "hasIssuesEnabled", "hasProjectsEnabled", "hasWikiEnabled",
    "homepageUrl", "isEmpty", "isFork", "mergeCommitAllowed", "visibility")
)
}
\arguments{
\item{repository}{Name of repository in \verb{[OWNER/]REPO} format.  \code{OWNER} defaults to name of authenticating user.}

\item{...}{Ignored}

\item{fields}{Data fields to include.  See \url{https://cli.github.com/manual/gh_repo_view}.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{gh_repo_view()} gets details of a GitHub repository
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_repo_view()
}
}
\seealso{
\url{https://cli.github.com/manual/gh_repo_view}
}
