% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_secret.r
\name{gh_secret_delete}
\alias{gh_secret_delete}
\title{Delete a secret in a GitHub repository}
\usage{
gh_secret_delete(secret_name, ..., repo = NULL)
}
\arguments{
\item{secret_name}{Name of secret.}

\item{...}{Ignored.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\description{
\code{gh_secret_delete()} deletes a secret in a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_secret_delete("SECRET_NAME")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_secret_delete}
}
