% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_secret.r
\name{gh_secret_list}
\alias{gh_secret_list}
\title{List secrets in a GitHub repository}
\usage{
gh_secret_list(..., repo = NULL)
}
\arguments{
\item{...}{Ignored.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
A data frame with labels with columns "name", "color", "description".
}
\description{
\code{gh_secret_list()} returns a data frame of secrets in a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_secret_list()
}
}
\seealso{
\url{https://cli.github.com/manual/gh_secret_list}
}
