% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_secret.r
\name{gh_secret_set}
\alias{gh_secret_set}
\title{Sets a secret in a GitHub repository}
\usage{
gh_secret_set(secret_name, body, ..., repo = NULL)
}
\arguments{
\item{secret_name}{Name of the secret.}

\item{body}{Value of the secret}

\item{...}{Ignored.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\description{
\code{gh_secret_set()} sets a secret in a GitHub repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_secret_set("SECRET_NAME", "SECRET_VALUE")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_secret_set}
}
