% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_label.r
\name{gh_label_edit}
\alias{gh_label_edit}
\title{Update label in a repository}
\usage{
gh_label_edit(
  what,
  ...,
  color = NULL,
  description = NULL,
  name = NULL,
  repo = NULL
)
}
\arguments{
\item{what}{A character string, integer, or \code{NULL}
\itemize{
\item \code{NULL} opens up repository home page
\item "commit" opens up last commit
\item "projects" opens repository projects
\item "settings" opens repository settings
\item "wiki" opens repository wiki
\item INT or "INT" opens issue or pull request INT
\item "PATH" opens file page of file with path PATH
\item "PATH:INT" opens file page of file with path PATH at line INT
\item "SHA" opens commit page of commit with sha SHA
}}

\item{...}{Ignored.}

\item{color}{Color of label.}

\item{description}{Description of label.}

\item{name}{New name of the label.}

\item{repo}{A string of another repository in \verb{[HOST/]OWNER/REPO} format.}
}
\value{
\code{NULL} invisibly
}
\description{
\code{gh_label_edit()} updates a label in a repository.
}
\examples{
\dontrun{
  # requires `gh` installed and authenticated and working directory in Github repository
  gh_label_edit("question", color = "orange")
}
}
\seealso{
\url{https://cli.github.com/manual/gh_label_edit}
}
